requirements.txt
absl-py==1.4.0
accelerate==1.0.1
annotated-types==0.7.0
anyio==4.5.2
attrs==23.1.0
av==12.3.0
boto3==1.28.36
botocore==1.31.36
brotli==1.2.0
cachetools==5.3.1
cairocffi==1.7.1
certifi
cffi==1.17.1
charset-normalizer==3.4.4
click==8.1.7
cloudpickle==1.6.0
contourpy==1.1.0
cycler==0.11.0
distro==1.9.0
einops==0.8.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastdtw==0.3.4
filelock==3.12.3
fire==0.7.1
fonttools==4.42.1
fsspec==2023.6.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
gym==0.17.3
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.11
igraph==0.10.6
imageio==2.31.2
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==6.0.1
inquirerpy==0.3.4
Jinja2==3.1.6
jiter==0.9.1
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.40.1
lmdb==1.4.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
modelscope==1.29.1
mpmath==1.3.0
msgpack==0.6.2
msgpack-numpy==0.4.8
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
mutagen==1.47.0
natsort==8.4.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
oauthlib==3.2.2
openai==1.98.0
opencv-contrib-python==4.8.0.76
packaging==23.1
pandas==2.0.3
pfzy==0.3.4
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.51
protobuf==4.24.2
psutil==7.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pyglet==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-transformers==1.2.0
pytz==2023.3
PyYAML==6.0.1
qwen-vl-utils==0.0.11
randomname==0.2.1
regex==2023.8.8
reportlab==4.4.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.2
sacremoses==0.0.53
safetensors==0.5.3
scipy==1.10.1
seaborn==0.12.2
sentencepiece==0.1.99
sentry-sdk==2.37.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
termcolor==2.4.0
texttable==1.6.7
timm==1.0.19
tokenizers==0.20.3
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==4.5.3
tqdm==4.66.1
transformers==4.46.3
triton==3.0.0
typing_extensions==4.13.2
tzdata==2023.3
urllib3==1.26.20
wandb==0.21.3
wcwidth==0.2.13
websockets==13.1
Werkzeug==2.3.7
yacs==0.1.8
zipp==3.16.2
